+++ package.json
@@ -12,7 +12,9 @@
     "react": "^17.0.2",
     "react-bootstrap": "^1.6.1",
     "react-dom": "^17.0.2",
+    "react-redux": "^7.2.4",
     "react-scripts": "4.0.3",
+    "redux": "^4.1.0",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
