--- src/Header.css
@@ -1,38 +1,0 @@
-.header {
-    width: 100%;
-    height: 50px;
-    background-color: #37b6ce;
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    /*#FFC996
-    #FF9D40*/
-}
-.logo {
-    //font-family: "Meiryo UI", serif;
-    font-family: Impact, serif;
-    font-size: 25px;
-    color: white;
-    margin-left: 10%;
-    padding: 1px 2px;
-    background-color: #ff9d40;
-    border-radius: 3px;
-}
-.search-bar {
-    margin-left: 10%;
-    width: calc(50% - 16px);
-    height: 25px;
-
-}
-.log-in-button {
-    margin-left: 5%;
-    height: 30px;
-    width: 5%;
-    background-color: #ffc996;
-}
-.sign-up-button {
-    margin-left: 1%;
-    height: 30px;
-    width: 5%;
-    background-color: #ff9d40;
-}
