+++ src/components/App.js
@@ -1,4 +1,4 @@
-import './App.css';
+import '../styles/App.css';
 import React from 'react';
 import Header from "./Header";
 import LeftPanel from "./LeftPanel";
