+++ src/components/LogInButton.js
@@ -1,0 +1,13 @@
+import React from "react";
+import {Button} from "antd";
+import "../styles/LogInButton.css"
+
+class LogInButton extends React.Component {
+    render() {
+        return (
+            <Button className="log-in-button">Log in</Button>
+        );
+    }
+}
+
+export default LogInButton;
