+++ src/components/Logo.js
@@ -1,0 +1,12 @@
+import React from "react";
+import "../styles/Logo.css";
+
+class Logo extends React.Component{
+    render() {
+        return (
+            <div id="logo-id" className="logo">SNPT</div>
+        );
+    }
+}
+
+export default Logo;
