+++ src/styles/FeedPanel.css
