+++ src/styles/Header.css
@@ -1,0 +1,8 @@
+.header {
+    width: 100%;
+    height: 50px;
+    background-color: #37b6ce;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+}
