+++ src/styles/LeftPanel.css
