+++ src/styles/LogInButton.css
@@ -1,0 +1,6 @@
+.log-in-button {
+    margin-left: 5%;
+    height: 30px;
+    width: 5%;
+    background-color: #ffc996;
+}
