+++ src/styles/RightPanel.css
