+++ src/styles/SearchBar.css
@@ -1,0 +1,5 @@
+.search-bar {
+    margin-left: 10%;
+    width: calc(50% - 16px);
+    height: 25px;
+}
