+++ src/styles/SignUpButton.css
@@ -1,0 +1,6 @@
+.sign-up-button {
+    margin-left: 1%;
+    height: 30px;
+    width: 5%;
+    background-color: #ff9d40;
+}
